Migrate string spans to be a dtype "STRING_SPAN", and a act as a leaf. (#15)
String spans now have a STRING_SPAN dtype, and contain {start: int32, end: int32} with a reference column that points to where the slice should be obtained. STRING_SPANs themselves act as leafs so they are analogous to a string.

This PR also brings back the old schema_test.py which we forgot to migrate.