Add OpenAPI => TS web client REST APIs. (#17)
We now are using OpenAPI to generate a TypeScript fetch client. See:
https://fastapi.tiangolo.com/advanced/generate-clients/

FastAPI serves the spec automatically from /openapi.json

When the dev server runs we now
1) run the typescript generation once by programmatically calling python
to generate an openapi.json file.
2) Listen for py changes and regenerate the typescript, with
watchdog[watchmedo] which is google's python nodemon (which I couldn't
get to watch python files).

Code is generated to `src/web/fast_api_client` which is *outside*
`src/web/src/...*.ts` to make it clear which code is generated.

<img width="270" alt="image"
src="https://user-images.githubusercontent.com/1100749/229658353-a38150c5-ac24-4ca5-98e2-27270127f5e1.png">

We check in the TypeScript files so we are able to see what the API
looks like from GitHub, and to watch diffs.

We also add 2 endpoints for listing sources, and listing a source's
fields.