Make the "Add filter" menu work (#40)
When the user clicks "Add filter" , we autocomplete with all the feature
names. Then the user selects a feature, and we autocomplete with all the
feature values (if there are not too many).

We don't yet have filters added to the global state. That will be done
in a follow up.


https://user-images.githubusercontent.com/2294279/231323286-362c54f8-05cf-4b66-99ac-b015589b4fe2.mp4