Add simple "activeConcept" selection from the UI. (#77)
I took a bunch of shortcuts to get here, which I will follow up with fixes for, but this unblocks us.

To activate a concept, you must go to "Edit Concept", select concept, embedding, and column.
The column will only be enabled if you have computed an embedding over it, that matches the embedding when you're selecting a concept.  For now, the other columns are disabled. In a follow up, I'll have them let you compute the embedding right from there.

UI fixes:
- When running an embedding index, close the menu and open the tasks menu. Add wiring for redux state for task menu being open.

Python fixes:
- Swap the constructor validation for Embeddings, Signal, Sources to be a regular validator to play nicely when the subclass 

https://user-images.githubusercontent.com/1100749/233517441-d1dd1de1-23a7-48f3-b896-2afe6aa87853.mov