Move hooks out of store (#107)
Moving the current hooks mostly as is out of the store file and into
individual files, making it easier to understand the store.ts file. In
addtion, changing the paramaters of the hooks to be an object so the
long list of parameters are named.

Adding simple testing to some of the hooks