Separate input_type and compute_type on signals. Fix embedding bug with outer dims of 1. (#225)
`input_type` refers to what we populate the UI with, `compute_type` is what compute() should be called with. If we are chaining signals, it's the ultimate input type (e.g. for TextEmbeddingModelSignals it's `input_type=text`, `compute_type=embedding`)

This fixes the UI dropdown:
<img width="706" alt="image" src="https://github.com/lilacai/lilac/assets/1100749/55e161ea-7607-4c19-8a38-69e1595cc87e">


Also fix a bug with embeddings where if embedding signals return a vector with an outer dim = 1, we incorrectly stack the embedding matrix. Now we squeeze, and yell if it's still not a vector.

We have a toxicity score on a UDF in the UI!

<img width="710" alt="image" src="https://github.com/lilacai/lilac/assets/1100749/47851b8b-117c-4dec-8008-29635bd0c083">

Fixes https://github.com/lilacai/lilac/issues/217