Add implicit negatives (#319)
Add implicit negatives when making concept models. A concept model can
now be tied to a specific dataset, or used in a dataset-agnostic way
(previous behavior).

- Rename `ConceptModelManager` to `ConceptModel`
- Rename `ConceptModel` to `LogisticEmbeddingModel`, which only
implements `score_embeddings()`
- `ConceptModel` can now take optional `dataset_info`. When
`dataset_info` is provided, we will randomly sample a bunch of negatives
from the dataset. Otherwise, we won't sample any negatives.
- Move the logic for randomly sampling sentences to `ConceptModel` since
it's internal details of that model. This means that we had to break a
circular dep or `dataset.py` depending on `ConceptModel`.