Fix routing to `signal.vector_compute_topk` (#351)
We regressed in routing to `signal.vector_compute_topk` after the
search_udf change. Also add detailed coverage for that codepath.
Also replace `flatten()` with `reshape(-1)` as flatten copies the data.

In a follow-up we will flip the entire logic in select_rows to first run
the topk, then duckdb which should result in 10x+ speedups.