Move "visibleFields" to the dataset store. (#366)
We should slowly migrate away from the utils that take stores and return values. These are very convenient but cause extremely weird bugs in svelte with re-renders.

If we want things to pre-computed, just precomputed them in the store and then derive state from them.

This fixes 2 bugs: 
- when a signal is computed, the fields don't react
- after a semantic search values were stale

In a follow up I'll continue to migrate stuff in view_utils.ts.