Store the current state in the URL. (#478)
Serialize the current state to the URL 

Each time the dataset store is updated, we will update the URL so it can
be saved and shared.

For instance:

https://lilacai-nikhil-staging.hf.space/datasets#local/spotify&query=%7B%22searches%22%3A%5B%7B%22path%22%3A%5B%22text%22%5D%2C%22query%22%3A%7B%22type%22%3A%22concept%22%2C%22concept_namespace%22%3A%22lilac%22%2C%22concept_name%22%3A%22profanity%22%2C%22embedding%22%3A%22sbert%22%7D%7D%5D%7D

Details:
- The global app store deals with all URL changes. Clients can subscribe
to changes. The post-hash value is always #identifier&state.
- The app store now exposes `persistedHashStore` which can be used to
add a store to the URL. This will subscribe to store changes and update
the URL. It will also subscribe to URL changes, and update the store.
- Currently only the dataset view uses this, but it will be usable by
concepts if we decide to have deeper links in concepts.

UI:
- Remove "reset view". Now we have a copy link button.
- To "reset the view", you can now click on the dataset icon in the
top-left.