Add a project_path to lilac_start. (#621)
`lilac start ~/my_project` will now start a project in that folder,
create a root lilac.yml, and start the webserver. When the server boots
up, we will call `load` which will show the tasks to load the datasets
from the lilac.yml.

Impl details:
- When passing a project path, we will check for lilac.yml. If it
doesn't exist, create it. This is now the project directory.

NOTE: The huggingface demos don't yet call the load on boot because
there is no lilac.yml file. This is because we call `gunicorn
server:app` directly, bypassing our public API of start_server. I'll
clean this up in a follow up.

Nothing changes in demo:
https://huggingface.co/spaces/lilacai/nikhil_staging