Update the global project config during changes. (#631)
High-level: lilac.yml now will be updated any time an action is
performed on a dataset.

Details:
- Datasets call into methods in `project` which will write to lilac.yml.
- The old config.yml is now gone, however for backwards compat I made
sure to have lilac.yml update from dataset configs if they exist.
- `lilac.yml` is written *before* computation. The manifest will reflect
the post-computation state. This means that when the server boots up, it
will try to call load on the fresh lilac.yml.