Add the python API for adding labels. (#667)
New API on `Dataset` to add a label to a row or set of rows.
```py
dataset.add_label(
    label_name: str,
    label_value: str,
    searches: Optional[Sequence[Search]] = None,
    filters: Optional[Sequence[FilterLike]] = None) -> None:
```

When adding a label, labels are written to `{label_name}.label.jsonl`
which contains json lines of the labels. This is not optimal, but we can
switch out the implementation later.

This then acts as a regular feature, with `(label_name, 'label')` to
access the actual label value. We also expose other metadata like
`(label_metadata, 'created')` which tells you the timestamp when the
label was created.

This also adds the [`freezegun`](https://github.com/spulec/freezegun)
dev-dependency for freezing time in python tests. `freezegun` is a very
popular library since python doesn't allow mocking datetime.now().