Use iterables for compute_signal and compute_embedding. (#706)
High-level:
- compute_signal() and compute_embedding() now directly call duckdb to
get source values. We use `fetch_df_chunk` to iteratively get chunks of
the response of the query.
- Downstream, everything is now iterable, which means all of the batch &
dataset utils are now generators.

This is the first step towards
https://github.com/lilacai/lilac/issues/660

A follow up will actually write embeddings iteratively.

Regression: now we do not support running signals over span fields. This
isn't really supported anywhere, but easy to fix if we decide we want it
later.

https://lilacai-nikhil-staging.hf.space/