Rework the huggingface deploy script: part 1 (#720)
- Move all the docker related images into a data-directory
(`lilac/hf_docker`) so it can be used inside the pip package. This
allows us to greatly shorten the code that is uploaded on the image, and
version all the files related to the image (`.Dockerfile`,
`docker_start.sh`, etc).
- Migrate all the uploads to use the commit API so it's atomatically
changed, and speeds up deployment.
- Remove .env and .env.demo from the image, everything will happen from
the environment flags via the UI.
- Make the huggingface_hub an optional dependency so we don't include it
in our pip package deps.

This is just part 1. Part 2 will improve the deploy_hf API so it can be
used via our API (and moved into the pip package).

https://huggingface.co/spaces/lilacai/nikhil_staging