Continue improving new chunker algorithm (#791)
The embedding calculation turned out to be pretty slow - it runs on
every sentence, which overall took ~7x longer than calculating
embeddings on the final chunks. So the total runtime for computing
embeddings got 8x slower.

Instead of using the general purpose embedders, I'm repalcing it with a
super cheap and simple bag of words embedding, computed over openai's
tokenizer tokens. The new embedder runs ~30x faster than even GTE-tiny,
so the total runtime for computing embeddings is only ~5-10% slower than
before.

The BoW vector is folded into a 128-vector, which is plain and simple
for now. I haven't gotten the UI working yet so I don't know yet how
much degradation we're getting in the chunking quality. The vector
dimension is an obvious place to fine-tune the approach later on if we
need to. The BoW embedding should at least pick up on things like code
vs. text vs. instructions.