Add parallelism to dataset.map with dask. (#847)
dataset.map() now takes a num_jobs argument for splitting the work. When
passed -1, it will use all the cores on the machine.

- We now automatically create a set of tasks for the user when calling
dataset.map(), showing up in the UI automatically. If you ask for N
jobs, N tasks are created in the UI. To avoid over-generalization, I
decided not to push this into tasks for now.
- When work is split between multiple jobs, each worker gets chunk
offsets and requests the data independently, and writes to shards
independently. The scheduler is responsible for joining all the jsonl
files to parquet at the very end.
- When you call map() with no output, we now write to disk to allow
communication between workers and the caller.
- Lots of unit tests to make sure continuation still works.


Other things I had to do:
- Add nest_asyncio, which patches asyncio to have event loops allowed to
be created in a nested fasion, which enables all the asyncio stuff from
jyupter notebooks.
- Lots of fixes to tasks.py to make it propagate errors.
- Adds a TestDaskLogger for use in unit tests where we want to
communicate between workers and the scheduler without dealing with dask
api.
- Remove freeze_time from tests that interact with dask. This completely
freezes the unit tests.
- PII tests seemed to have been checked in in half-baked state, since we
have confidence scores but dont assert them in unit tests.