Allow custom signals to work with dask processes. (#880)
NOTE: This does not fix Jyupter notebooks which cannot pickle classes
defined in it.

We do this by resolving the signal in the scheduler process, and
allowing pickle to pickle the entire class, instead of trying to look it
up in the registry when the dask worker boots up.

- Upgrade pydantic. New versions support pickling.
- Fix the issue where we swallow errors with "signal outputted a
different number of elements than input". Now we separate these and
throw if there is a real error, printing the full traceback.