Add threaded task execution. (#920)
Adds a `type='threads'|'processes'` param to map() to determine whether
to split jobs amongst threads or processes.

When a task type is threads, we create a ThreadPoolExecutor with the
number of tasks submitted. This then spins up threads, and executes
them.

To do this, I simulated the logging API of dask for threads.